fix: password field type
