+++ packages/shared/src/components/fields/PasswordField.spec.tsx
@@ -1,0 +1,88 @@
+import React from 'react';
+import { render, RenderResult, screen, waitFor } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
+import { TextField, TextFieldProps } from './TextField';
+import { PasswordField } from './PasswordField';
+
+const renderComponent = (props: Partial<TextFieldProps> = {}): RenderResult => {
+  const defaultProps: TextFieldProps = {
+    inputId: 'name',
+    name: 'name',
+    label: 'Name',
+  };
+  return render(<PasswordField {...defaultProps} {...props} />);
+};
+
+const getInput = (): HTMLInputElement =>
+  screen.getByRole('textbox') as HTMLInputElement;
+
+it('should render a password field', async () => {
+  renderComponent({
+    fieldType: 'tertiary',
+    placeholder: 'Create a password',
+    label: 'Create a password',
+    type: 'password',
+    role: 'textbox',
+  });
+  const input = getInput();
+  await waitFor(() => expect(input.type).toEqual('password'));
+});
+
+it('should show password strength on input', async () => {
+  renderComponent({
+    fieldType: 'tertiary',
+    placeholder: 'Create a password',
+    label: 'Create a password',
+    type: 'password',
+    role: 'textbox',
+  });
+  const input = getInput();
+  input.value = 'a';
+  input.dispatchEvent(new Event('input', { bubbles: true }));
+  const el = screen.getByText('Risky');
+  await waitFor(() => expect(el).toHaveClass('text-theme-status-error'));
+});
+
+it('should show medium password strength on input', async () => {
+  renderComponent({
+    fieldType: 'tertiary',
+    placeholder: 'Create a password',
+    label: 'Create a password',
+    type: 'password',
+    role: 'textbox',
+  });
+  const input = getInput();
+  input.value = 'asAS12!@';
+  input.dispatchEvent(new Event('input', { bubbles: true }));
+  const el = screen.getByText(`You're almost there`);
+  await waitFor(() => expect(el).toHaveClass('text-theme-status-warning'));
+});
+
+it('should show strong password strength on input', async () => {
+  renderComponent({
+    fieldType: 'tertiary',
+    placeholder: 'Create a password',
+    label: 'Create a password',
+    type: 'password',
+    role: 'textbox',
+  });
+  const input = getInput();
+  input.value = 'asAS12!@as';
+  input.dispatchEvent(new Event('input', { bubbles: true }));
+  const el = screen.getByText('Strong as it gets');
+  await waitFor(() => expect(el).toHaveClass('text-theme-status-success'));
+});
+
+it('should show the password plain text on icon click', async () => {
+  renderComponent({
+    fieldType: 'tertiary',
+    placeholder: 'Create a password',
+    label: 'Create a password',
+    type: 'password',
+    role: 'textbox',
+  });
+  const button = screen.getByRole('button');
+  await button.click();
+  const input = getInput();
+  await waitFor(() => expect(input.type).toEqual('text'));
+});
