+++ packages/shared/src/components/fields/TextField.spec.tsx
@@ -75,74 +75,3 @@ it('should show label and placeholder based in its state in tertiary mode', asyn
   input.focus();
   await waitFor(() => expect(input.placeholder).toEqual('Placeholder'));
 });
-
-it('should render a password field', async () => {
-  renderComponent({
-    fieldType: 'tertiary',
-    placeholder: 'Create a password',
-    label: 'Create a password',
-    type: 'password',
-    role: 'textbox',
-  });
-  const input = getInput();
-  await waitFor(() => expect(input.type).toEqual('password'));
-});
-
-it('should show password strength on input', async () => {
-  renderComponent({
-    fieldType: 'tertiary',
-    placeholder: 'Create a password',
-    label: 'Create a password',
-    type: 'password',
-    role: 'textbox',
-  });
-  const input = getInput();
-  input.value = 'a';
-  input.dispatchEvent(new Event('input', { bubbles: true }));
-  const el = screen.getByText('Risky');
-  await waitFor(() => expect(el).toHaveClass('text-theme-status-error'));
-});
-
-it('should show medium password strength on input', async () => {
-  renderComponent({
-    fieldType: 'tertiary',
-    placeholder: 'Create a password',
-    label: 'Create a password',
-    type: 'password',
-    role: 'textbox',
-  });
-  const input = getInput();
-  input.value = 'asAS12!@';
-  input.dispatchEvent(new Event('input', { bubbles: true }));
-  const el = screen.getByText(`You're almost there`);
-  await waitFor(() => expect(el).toHaveClass('text-theme-status-warning'));
-});
-
-it('should show strong password strength on input', async () => {
-  renderComponent({
-    fieldType: 'tertiary',
-    placeholder: 'Create a password',
-    label: 'Create a password',
-    type: 'password',
-    role: 'textbox',
-  });
-  const input = getInput();
-  input.value = 'asAS12!@as';
-  input.dispatchEvent(new Event('input', { bubbles: true }));
-  const el = screen.getByText('Strong as it gets');
-  await waitFor(() => expect(el).toHaveClass('text-theme-status-success'));
-});
-
-it('should show the password plain text on icon click', async () => {
-  renderComponent({
-    fieldType: 'tertiary',
-    placeholder: 'Create a password',
-    label: 'Create a password',
-    type: 'password',
-    role: 'textbox',
-  });
-  const button = screen.getByRole('button');
-  await button.click();
-  const input = getInput();
-  await waitFor(() => expect(input.type).toEqual('text'));
-});
